--- cloud-batch-python/python-batch/simple-container.yaml
@@ -1,31 +1,0 @@
-#  Copyright 2023 Google LLC
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-project_id: "fluent-demo"
-region: "us-central1"
-
-job_prefix: 'container-'
-machine_type: "n2-standard-2"
-container: 
-  image_uri: "gcr.io/google-containers/busybox"
-  entry_point: "/bin/sh"
-  commands: ["-c", "echo Hello world! This is task ${BATCH_TASK_INDEX}. This job has a total of ${BATCH_TASK_COUNT} tasks."]
-
-task_count: 4 #optional
-parallelism: 4 #optional
-task_count_per_node: 2 #optional
-install_gpu_drivers: true
-
-labels: 
-  env: "pybatch"
-  type: "hello"
