+++ python-batch/nvidia-smi-container.yaml
@@ -14,23 +14,27 @@
 project_id: "project_id"
 region: "us-central1"
 
-job_prefix: 'deepvariant-'
-machine_type: "n1-standard-64"
-volumes:
-  - {bucket_name:  "demo-jrt", gcs_path:  "/mnt/disks/work"}
-  - {bucket_name:  "deepvariant", gcs_path:  "/mnt/disks/deepvariant"}
+task_count: 1 #optional
+parallelism: 1 #optional
 
-container: 
-  image_uri: "gcr.io/deepvariant-docker/deepvariant:1.2.0"
-  #image_uri: "gcr.io/google-containers/busybox"
-  entry_point: "/bin/sh"
-  commands: ["/mnt/disks/work/deepvariant.sh"]
-
+job_prefix: 'nvidia-smi-'
 task_count: 1 #optional
 parallelism: 1 #optional
 task_count_per_node: 1 #optional
 
+machine_type: "n1-standard-2"
+accelerator:
+ type: "nvidia-tesla-t4"
+ count: 1
+install_gpu_drivers: y
+cpu_milli: 1000 #optional
+memory_mib: 102400 #optional
+
+container: 
+  image_uri: "nvidia/cuda:12.1.1-devel-centos7" 
+  entry_point: "/bin/bash"
+  commands: ["-c", "nvidia-smi"]
 
 labels: 
-  env: "deepvariant"
-  type: "1_2_0"
+  env: "container"
+  type: "nvidia"
