+++ python-batch/quantum/cuquantum-noise.yaml
@@ -11,44 +11,36 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-project_id: "fluent-hpc-demo1"
+project_id: "project_id"
 region: "us-central1"
 
-job_prefix: 'build-'
-#template_link: "projects/fluent-hpc-demo1/global/instanceTemplates/fluent-template3"
-machine_type: "n1-standard-4"
+volumes:
+  - {bucket_name: "<my_bucket>", gcs_path:  "/mnt/disks/work"}
 
-nfs_server: "10.243.163.34"
-#nfs_path: "/mnt/nfs1"
-nfs_remote_path: "/nfsshare"
+job_prefix: 'cuquantum-noise-'
 
-bucket_name:  "fluent-hpc-demo1-bucket"
-#gcs_path: "/mnt/gcs1" #optional
+task_count: 100 #optional
+parallelism: 10 #optional
+task_count_per_node: 2 #optional
 
-task_count: 4 #optional
-parallelism: 2 #optional
-task_count_per_node: 1 #optional
-
+machine_type: "n1-standard-2"
 accelerator:
- type: "nvidia-tesla-v100"
+ type: "nvidia-tesla-t4"
  count: 1
 install_gpu_drivers: y
 cpu_milli: 1000 #optional
 memory_mib: 102400 #optional
 
+container: 
+  image_uri: "nvcr.io/nvidia/cuquantum-appliance:22.07-cirq" 
+  entry_point: "/bin/bash"
+  commands: 
+    - "-c"
+    - "
+      mkdir -p /mnt/disks/work/dir_${BATCH_JOB_ID}; 
+      python3 /mnt/disks/work/cuquantum-noise.py >> /mnt/disks/work/dir_${BATCH_JOB_ID}/out_${BATCH_TASK_INDEX}.txt
+      "
 
 labels: 
-  env: "fluent2"
-  type: "gpu-v100"
-
-# No comments allowed in script text. 
-
-script_text: | 
-  #!/bin/bash 
-  git clone https://github.com/NVIDIA/cuda-samples.git
-  cd cuda-samples/Samples/4_CUDA_Libraries/batchCUBLAS/
-  make
-  ./batchCUBLAS
-
-  sleep 1200
- 
+  env: "container"
+  type: "quantum"
